+++ package.py
@@ -140,6 +140,9 @@ def list_files(top_path, log=None):
     results = []
 
     for root, dirs, files in os.walk(top_path, followlinks=True):
+        # Sort directories and files to ensure they are always processed in the same order
+        dirs.sort()
+        files.sort()
         for file_name in files:
             file_path = os.path.join(root, file_name)
             relative_path = os.path.relpath(file_path, top_path)
@@ -211,6 +214,9 @@ def yesno_bool(val):
 
 def emit_dir_content(base_dir):
     for root, dirs, files in os.walk(base_dir, followlinks=True):
+        # Sort directories and files to ensure they are always processed in the same order
+        dirs.sort()
+        files.sort()
         if root != base_dir:
             yield os.path.normpath(root)
         for name in files:
@@ -596,6 +602,9 @@ class ZipContentFilter:
                 yield path
         else:
             for root, dirs, files in os.walk(path, followlinks=True):
+                # Sort directories and files to ensure they are always processed in the same order
+                dirs.sort()
+                files.sort()
                 o, d = norm_path(path, root)
                 # log.info('od: %s %s', o, d)
                 if root != path:
@@ -981,7 +990,7 @@ def docker_run_command(build_root, command, runtime,
                 '-e', 'SSH_AUTH_SOCK=/tmp/ssh_sock',
             ])
 
-    if platform.system() == 'Linux':
+    if platform.system() in ('Linux', 'Darwin'):
         if pip_cache_dir:
             pip_cache_dir = os.path.abspath(pip_cache_dir)
             docker_cmd.extend([